fe done

+++ database/migrations/2025_08_29_090124_create_sgds_table.php
@@ -10,31 +10,31 @@ return new class extends Migration
     {
         Schema::create('sgds', function (Blueprint $table) {
             $table->id();
-            $table->string('desa'); // Kampung Dalam
-            $table->string('kecamatan'); // Siak
-            $table->string('kabupaten'); // Siak
-            $table->string('icon')->nullable(); // Siak
-            $table->year('tahun'); // 2024
+            $table->string('desa');       // Kampung Dalam
+            $table->string('kecamatan');  // Siak
+            $table->string('kabupaten');  // Siak
+            $table->string('icon')->nullable();
+            $table->year('tahun');        // 2021 - sekarang
 
-            // Indikator SGDs
-            $table->integer('desa_tanpa_kemiskinan')->default(0);
-            $table->integer('desa_tanpa_kelaparan')->default(0);
-            $table->integer('desa_sehat_sejahtera')->default(0);
-            $table->integer('pendidikan_desa_berkualitas')->default(0);
-            $table->integer('keterlibatan_perempuan_desa')->default(0);
-            $table->integer('desa_layak_air_bersih')->default(0);
-            $table->integer('desa_berenergi_bersih')->default(0);
-            $table->integer('pertumbuhan_ekonomi_desa')->default(0);
-            $table->integer('infrastruktur_desa')->default(0);
-            $table->integer('ketimpangan_desa')->default(0);
-            $table->integer('kawasan_desa')->default(0);
-            $table->integer('konsumsi_produksi_desa')->default(0);
-            $table->integer('perubahan_iklim_desa')->default(0);
-            $table->integer('ekosistem_darat_desa')->default(0);
-            $table->integer('ekosistem_laut_desa')->default(0);
-            $table->integer('desa_damai')->default(0);
-            $table->integer('kemitraan_desa')->default(0);
-            $table->integer('kelembagaan_desa_dinamis')->default(0);
+            // Indikator SGDs (2 angka dibelakang koma)
+            $table->decimal('desa_tanpa_kemiskinan', 5, 2)->default(0);
+            $table->decimal('desa_tanpa_kelaparan', 5, 2)->default(0);
+            $table->decimal('desa_sehat_sejahtera', 5, 2)->default(0);
+            $table->decimal('pendidikan_desa_berkualitas', 5, 2)->default(0);
+            $table->decimal('keterlibatan_perempuan_desa', 5, 2)->default(0);
+            $table->decimal('desa_layak_air_bersih', 5, 2)->default(0);
+            $table->decimal('desa_berenergi_bersih', 5, 2)->default(0);
+            $table->decimal('pertumbuhan_ekonomi_desa', 5, 2)->default(0);
+            $table->decimal('infrastruktur_desa', 5, 2)->default(0);
+            $table->decimal('ketimpangan_desa', 5, 2)->default(0);
+            $table->decimal('kawasan_desa', 5, 2)->default(0);
+            $table->decimal('konsumsi_produksi_desa', 5, 2)->default(0);
+            $table->decimal('perubahan_iklim_desa', 5, 2)->default(0);
+            $table->decimal('ekosistem_darat_desa', 5, 2)->default(0);
+            $table->decimal('ekosistem_laut_desa', 5, 2)->default(0);
+            $table->decimal('desa_damai', 5, 2)->default(0);
+            $table->decimal('kemitraan_desa', 5, 2)->default(0);
+            $table->decimal('kelembagaan_desa_dinamis', 5, 2)->default(0);
 
             $table->timestamps();
         });
